Adapt tests to transcript size validation
Move some common test definitions to conftest.py
